refactor: move ApplierData's StateMachineApi impl next to its type
# Summary

`Change`, `OnChange` and the `StateMachineApi for ApplierData` impl move from
`sm_v003/sm_v003.rs` into `applier/applier_data`, where `ApplierData` is
defined. `applier` no longer depends on `sm_v003` in production code.

# Details

The impl sat in another module, a hundred lines away from the struct it
implements and interleaved with unrelated `SMV003` methods. Reading how an
applied entry reaches the state machine meant jumping between the two files,
and the split forced `applier` to import the change types from `sm_v003` while
`sm_v003` imported the applier back.

Only `applier` and `sm_v003` ever named `Change` and `OnChange`, so the
re-exports from `sm_v003` are dropped rather than forwarded. Callers hand
`set_on_change_applied` a boxed closure and never spell the alias.

The only remaining reference from `applier` to `sm_v003` is in `applier_test`,
which constructs a whole state machine on purpose.
